+++ examples/create-mysql-database/main.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     filess = {
-      source = "registry.terraform.io/filess/filess"
+      source = "app.terraform.io/filess/provider/filessdedicated"
       version = ">=1.0.0"
     }
   }